SquareGrid cellsize u32

+++ src/cli.rs
@@ -61,7 +61,7 @@ pub struct Cli {
     pub geometric_error_above_leaf: Option<f64>,
     /// Set the 2D cell size for the grid that is used for constructing the quadtree. In input units (eg. meters).
     #[arg(long, default_value = "250")]
-    pub grid_cellsize: Option<u16>,
+    pub grid_cellsize: Option<u32>,
     /// Generate the quadtree directly from a grid.tsv file, skipping the extent computation and feature indexing. A grid.tsv file is created with the --grid-export option. Used for debugging.
     #[arg(long)]
     pub grid_file: Option<String>,
